+++ src-tauri/improvie-command/src/command/mod.rs
@@ -1,4 +1,4 @@
 pub mod dialog;
 pub mod health_check;
 pub mod items;
-pub mod playlists;
+pub mod plays;
