+++ src-tauri/improvie-infra/src/model/mod.rs
@@ -1,3 +1,3 @@
 pub mod health_check;
 pub mod items;
-pub mod playlists;
+pub mod plays;
